+++ src/lib/model/Status.ts
@@ -3,5 +3,6 @@
  */
 export enum Status {
   ACTIVE = 'ACTIVE',
+  ARCHIVED = 'ARCHIVED',
   DELETED = 'DELETED',
 }